Froala enhancements
If the Froala package is added, a config var is added to the admin for
the license key. Also added a starting height of 300px as the editor
seemed very compressed upon init.

The `var key` was changed to `var name` because of a potential conflict
however I don't believe that was necessary and can be ignored or
reverted.