Move myAccount to dropdown
As discussed in the forums
(http://forums.orionjs.org/t/what-about-removing-my-account/65) I made
the following changes:

Bootstrap Package
1) I created a right dropdown for with responsive support by adding a
user icon in the top right corner when in mobile mode. (I also modified
the layout.less for this part)

2)I moved the each adminLinks ‘user’ helper cycle from the sidebar to
the dropdown. (As it already is in the materialize theme)

3)I added a logout event for the orionBootstrapHeader Template.

4)I added the myAccount.profile and myAccount.password links to the
‘user’ section

The Materialize theme doesn’t require any modify because it was already
using a right dropdown for user management.